# server/Records/openapi.json
{"openapi":"3.1.0","info":{"title":"record","version":"0.1.0"},"paths":{"/graphql/patient":{"get":{"summary":"Handle Http Get","operationId":"handle_http_get_graphql_patient_get","responses":{"200":{"description":"The GraphiQL integrated development environment.","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found if GraphiQL or query via GET are not enabled."}}},"post":{"summary":"Handle Http Post","operationId":"handle_http_post_graphql_patient_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/graphql/doctor":{"get":{"summary":"Handle Http Get","operationId":"handle_http_get_graphql_doctor_get","responses":{"200":{"description":"The GraphiQL integrated development environment.","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found if GraphiQL or query via GET are not enabled."}}},"post":{"summary":"Handle Http Post","operationId":"handle_http_post_graphql_doctor_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/records/get-patient-details":{"get":{"summary":"Get Patient Details","operationId":"get_patient_details_records_get_patient_details_get","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SelectPatient"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/doctor/recent-patients":{"get":{"summary":"Get Recent Patients","operationId":"get_recent_patients_records_doctor_recent_patients_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/records/prescription/symptoms-signs":{"get":{"summary":"Get Prescriptions","operationId":"get_prescriptions_records_prescription_symptoms_signs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/records/prescription/diagnosis":{"get":{"summary":"Get Prescriptions","operationId":"get_prescriptions_records_prescription_diagnosis_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/records/notes":{"get":{"summary":"Get Prescriptions","operationId":"get_prescriptions_records_notes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Get Prescriptions","operationId":"get_prescriptions_records_notes_post","parameters":[{"name":"note","in":"query","required":true,"schema":{"type":"string","title":"Note"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/upload/{type}":{"post":{"summary":"Upload Report","operationId":"upload_report_records_upload__type__post","parameters":[{"name":"type","in":"path","required":true,"schema":{"type":"string","title":"Type"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseReportTemplate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/select-patient":{"post":{"summary":"Verify Doctor","operationId":"verify_doctor_records_select_patient_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SelectPatient"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/verify-request":{"post":{"summary":"Verify Doctor Request","operationId":"verify_doctor_request_records_verify_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTP"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/add-prescription":{"post":{"summary":"Add Prescriptions","operationId":"add_prescriptions_records_add_prescription_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SymptomsSignsDiagnosis"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BaseMetaTemplate":{"properties":{"patientNIC":{"type":"string","title":"Patientnic"},"date":{"type":"string","title":"Date"},"time":{"type":"string","title":"Time"},"practitioner":{"type":"string","title":"Practitioner"},"clinic":{"type":"string","title":"Clinic"},"recorder":{"type":"string","title":"Recorder"},"instructions":{"type":"string","title":"Instructions"}},"type":"object","required":["patientNIC","date","time","practitioner","clinic","recorder","instructions"],"title":"BaseMetaTemplate"},"BaseReportTemplate":{"properties":{"mata":{"$ref":"#/components/schemas/BaseMetaTemplate"},"results":{"anyOf":[{"$ref":"#/components/schemas/CBCReportTemplate"},{"$ref":"#/components/schemas/UFRTemplate"},{"$ref":"#/components/schemas/LFTReportTemplate"},{"$ref":"#/components/schemas/CRPReportTemplate"},{"$ref":"#/components/schemas/FBSReportTemplate"},{"$ref":"#/components/schemas/SerumCreatinineReportTemplate"},{"$ref":"#/components/schemas/SerumElectrolytesReportTemplate"},{"$ref":"#/components/schemas/LipidProfileReportTemplate"},{"$ref":"#/components/schemas/HbA1cReportTemplate"},{"$ref":"#/components/schemas/ESRReportTemplate"},{"$ref":"#/components/schemas/TFTReportTemplate"}],"title":"Results"}},"type":"object","required":["mata","results"],"title":"BaseReportTemplate"},"CBCReportTemplate":{"properties":{"hemoglobin":{"type":"string","title":"Hemoglobin","metadata":{"loinc_code":"718-7","official_name":"Hemoglobin","unit":"g/dL"}},"totalLeukocyteCount":{"type":"string","title":"Totalleukocytecount","metadata":{"loinc_code":"6690-2","official_name":"Leukocytes","unit":"10^3/µL"}},"neutrophils":{"type":"string","title":"Neutrophils","metadata":{"loinc_code":"770-8","official_name":"Neutrophils","unit":"%"}},"lymphocytes":{"type":"string","title":"Lymphocytes","metadata":{"loinc_code":"731-0","official_name":"Lymphocytes","unit":"%"}},"eosinophils":{"type":"string","title":"Eosinophils","metadata":{"loinc_code":"711-2","official_name":"Eosinophils","unit":"%"}},"monocytes":{"type":"string","title":"Monocytes","metadata":{"loinc_code":"742-7","official_name":"Monocytes","unit":"%"}},"basophils":{"type":"string","title":"Basophils","metadata":{"loinc_code":"706-2","official_name":"Basophils","unit":"%"}},"plateletCount":{"type":"string","title":"Plateletcount","metadata":{"loinc_code":"777-3","official_name":"Platelet Count","unit":"10^3/µL"}},"totalRBCCount":{"type":"string","title":"Totalrbccount","metadata":{"loinc_code":"789-8","official_name":"Red Blood Cell Count","unit":"10^6/µL"}},"hematocrit":{"type":"string","title":"Hematocrit","metadata":{"loinc_code":"4544-3","official_name":"Hematocrit","unit":"%"}},"meanCorpuscularVolume":{"type":"string","title":"Meancorpuscularvolume","metadata":{"loinc_code":"787-2","official_name":"Mean Corpuscular Volume","unit":"fL"}},"meanCellHemoglobin":{"type":"string","title":"Meancellhemoglobin","metadata":{"loinc_code":"785-6","official_name":"Mean Corpuscular Hemoglobin","unit":"pg"}},"meanCellHemoglobinConcentration":{"type":"string","title":"Meancellhemoglobinconcentration","metadata":{"loinc_code":"786-4","official_name":"Mean Corpuscular Hemoglobin Concentration","unit":"g/dL"}}},"type":"object","required":["hemoglobin","totalLeukocyteCount","neutrophils","lymphocytes","eosinophils","monocytes","basophils","plateletCount","totalRBCCount","hematocrit","meanCorpuscularVolume","meanCellHemoglobin","meanCellHemoglobinConcentration"],"title":"CBCReportTemplate"},"CRPReportTemplate":{"properties":{"crpLevel":{"type":"string","title":"Crplevel","metadata":{"loinc_code":"1988-5","official_name":"C-Reactive Protein","unit":"mg/L"}}},"type":"object","required":["crpLevel"],"title":"CRPReportTemplate"},"ESRReportTemplate":{"properties":{"esr":{"type":"string","title":"Esr","metadata":{"loinc_code":"4537-8","official_name":"Erythrocyte Sedimentation Rate","unit":"mm/hr"}}},"type":"object","required":["esr"],"title":"ESRReportTemplate"},"FBSReportTemplate":{"properties":{"fastingBloodSugar":{"type":"string","title":"Fastingbloodsugar","metadata":{"loinc_code":"1558-6","official_name":"Glucose [Bld] Fasting","unit":"mg/dL"}}},"type":"object","required":["fastingBloodSugar"],"title":"FBSReportTemplate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HbA1cReportTemplate":{"properties":{"hba1c":{"type":"string","title":"Hba1C","metadata":{"loinc_code":"4548-4","official_name":"Hemoglobin A1c","unit":"%"}},"estimatedAvgGlucose":{"type":"string","title":"Estimatedavgglucose","metadata":{"loinc_code":"1557-8","official_name":"Estimated Average Glucose","unit":"mg/dL"}}},"type":"object","required":["hba1c","estimatedAvgGlucose"],"title":"HbA1cReportTemplate"},"Indications":{"type":"string","enum":["symptoms","signs","diagnosis"],"title":"Indications"},"LFTReportTemplate":{"properties":{"serumBilirubinTotal":{"type":"string","title":"Serumbilirubintotal","metadata":{"loinc_code":"1975-2","official_name":"Bilirubin Total","unit":"mg/dL"}},"serumBilirubinDirect":{"type":"string","title":"Serumbilirubindirect","metadata":{"loinc_code":"1968-7","official_name":"Bilirubin Direct","unit":"mg/dL"}},"serumBilirubinIndirect":{"type":"string","title":"Serumbilirubinindirect","metadata":{"loinc_code":"1963-8","official_name":"Bilirubin Indirect","unit":"mg/dL"}},"sGPTALT":{"type":"string","title":"Sgptalt","metadata":{"loinc_code":"1742-6","official_name":"ALT","unit":"U/L"}},"sGOTAST":{"type":"string","title":"Sgotast","metadata":{"loinc_code":"1920-8","official_name":"AST","unit":"U/L"}},"serumAlkalinePhosphatase":{"type":"string","title":"Serumalkalinephosphatase","metadata":{"loinc_code":"6768-6","official_name":"Alkaline Phosphatase","unit":"U/L"}},"serumProtein":{"type":"string","title":"Serumprotein","metadata":{"loinc_code":"2885-2","official_name":"Total Protein","unit":"g/dL"}},"serumAlbumin":{"type":"string","title":"Serumalbumin","metadata":{"loinc_code":"1751-7","official_name":"Albumin","unit":"g/dL"}},"globulin":{"type":"string","title":"Globulin","metadata":{"loinc_code":"2336-8","official_name":"Globulin","unit":"g/dL"}},"agRatio":{"type":"string","title":"Agratio","metadata":{"loinc_code":"2339-2","official_name":"Albumin/Globulin Ratio","unit":"ratio"}}},"type":"object","required":["serumBilirubinTotal","serumBilirubinDirect","serumBilirubinIndirect","sGPTALT","sGOTAST","serumAlkalinePhosphatase","serumProtein","serumAlbumin","globulin","agRatio"],"title":"LFTReportTemplate"},"LipidProfileReportTemplate":{"properties":{"totalCholesterol":{"type":"string","title":"Totalcholesterol","metadata":{"loinc_code":"2093-3","official_name":"Cholesterol Total","unit":"mg/dL"}},"triglycerides":{"type":"string","title":"Triglycerides","metadata":{"loinc_code":"2571-8","official_name":"Triglycerides","unit":"mg/dL"}},"hdl":{"type":"string","title":"Hdl","metadata":{"loinc_code":"2085-9","official_name":"HDL Cholesterol","unit":"mg/dL"}},"ldl":{"type":"string","title":"Ldl","metadata":{"loinc_code":"2089-1","official_name":"LDL Cholesterol","unit":"mg/dL"}},"vldl":{"type":"string","title":"Vldl","metadata":{"loinc_code":"3043-7","official_name":"VLDL Cholesterol","unit":"mg/dL"}},"ldlToHdlRatio":{"type":"string","title":"Ldltohdlratio","metadata":{"loinc_code":"9830-1","official_name":"LDL/HDL Ratio","unit":"ratio"}},"totalCholesterolToHdlRatio":{"type":"string","title":"Totalcholesteroltohdlratio","metadata":{"loinc_code":"9832-7","official_name":"Cholesterol/HDL Ratio","unit":"ratio"}},"tgToHdlRatio":{"type":"string","title":"Tgtohdlratio","metadata":{"loinc_code":"50010-2","official_name":"Triglyceride/HDL Ratio","unit":"ratio"}},"nonHdlCholesterol":{"type":"string","title":"Nonhdlcholesterol","metadata":{"loinc_code":"43396-1","official_name":"Non-HDL Cholesterol","unit":"mg/dL"}}},"type":"object","required":["totalCholesterol","triglycerides","hdl","ldl","vldl","ldlToHdlRatio","totalCholesterolToHdlRatio","tgToHdlRatio","nonHdlCholesterol"],"title":"LipidProfileReportTemplate"},"MedicationUploadModel":{"properties":{"display":{"type":"string","title":"Display"},"frequency":{"type":"string","title":"Frequency"},"mealTiming":{"type":"string","title":"Mealtiming"},"dosage":{"type":"string","title":"Dosage"},"route":{"type":"string","title":"Route"},"prescriber":{"type":"string","title":"Prescriber"},"meta":{"type":"object","title":"Meta"}},"type":"object","required":["display","frequency","mealTiming","dosage","route","prescriber","meta"],"title":"MedicationUploadModel"},"OTP":{"properties":{"otp":{"type":"string","title":"Otp"}},"type":"object","required":["otp"],"title":"OTP"},"SelectPatient":{"properties":{"NIC":{"type":"string","title":"Nic"}},"type":"object","required":["NIC"],"title":"SelectPatient"},"SerumCreatinineReportTemplate":{"properties":{"serumCreatinine":{"type":"string","title":"Serumcreatinine","metadata":{"loinc_code":"2160-0","official_name":"Creatinine","unit":"mg/dL"}}},"type":"object","required":["serumCreatinine"],"title":"SerumCreatinineReportTemplate"},"SerumElectrolytesReportTemplate":{"properties":{"sodium":{"type":"string","title":"Sodium","metadata":{"loinc_code":"2951-2","official_name":"Sodium","unit":"mmol/L"}},"potassium":{"type":"string","title":"Potassium","metadata":{"loinc_code":"2823-3","official_name":"Potassium","unit":"mmol/L"}},"chloride":{"type":"string","title":"Chloride","metadata":{"loinc_code":"2075-0","official_name":"Chloride","unit":"mmol/L"}},"bicarbonate":{"type":"string","title":"Bicarbonate","metadata":{"loinc_code":"1963-8","official_name":"Bicarbonate","unit":"mmol/L"}},"calcium":{"type":"string","title":"Calcium","metadata":{"loinc_code":"17861-6","official_name":"Calcium","unit":"mg/dL"}},"magnesium":{"type":"string","title":"Magnesium","metadata":{"loinc_code":"19123-9","official_name":"Magnesium","unit":"mg/dL"}}},"type":"object","required":["sodium","potassium","chloride","bicarbonate","calcium","magnesium"],"title":"SerumElectrolytesReportTemplate"},"SymptomsSignsDiagnosis":{"properties":{"timeStamp":{"type":"string","format":"date-time","title":"Timestamp"},"indications":{"$ref":"#/components/schemas/Indications"},"doctorName":{"type":"string","title":"Doctorname"},"diagnosedCondition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Diagnosedcondition"},"medications":{"items":{"$ref":"#/components/schemas/MedicationUploadModel"},"type":"array","title":"Medications"},"note":{"type":"string","title":"Note"}},"type":"object","required":["timeStamp","indications","doctorName","medications","note"],"title":"SymptomsSignsDiagnosis"},"TFTReportTemplate":{"properties":{"tsh":{"type":"string","title":"Tsh","metadata":{"loinc_code":"3016-3","official_name":"Thyroid Stimulating Hormone","unit":"µIU/mL"}},"t3":{"type":"string","title":"T3","metadata":{"loinc_code":"3054-0","official_name":"Triiodothyronine (T3)","unit":"ng/dL"}},"t4":{"type":"string","title":"T4","metadata":{"loinc_code":"3026-2","official_name":"Thyroxine (T4)","unit":"µg/dL"}}},"type":"object","required":["tsh","t3","t4"],"title":"TFTReportTemplate"},"UFRTemplate":{"properties":{"quantity":{"type":"string","title":"Quantity","metadata":{"loinc_code":"5767-9","official_name":"Quantity","unit":"mL"}},"color":{"type":"string","title":"Color","metadata":{"loinc_code":"5778-6","official_name":"Color of Urine","unit":"N/A"}},"transparency":{"type":"string","title":"Transparency","metadata":{"loinc_code":"5780-2","official_name":"Urine Clarity","unit":"N/A"}},"specificGravity":{"type":"string","title":"Specificgravity","metadata":{"loinc_code":"5811-5","official_name":"Specific Gravity","unit":"ratio"}},"pH":{"type":"string","title":"Ph","metadata":{"loinc_code":"2756-5","official_name":"Urine pH","unit":"pH"}},"protein":{"type":"string","title":"Protein","metadata":{"loinc_code":"5804-0","official_name":"Urine Protein","unit":"mg/dL"}},"sugar":{"type":"string","title":"Sugar","metadata":{"loinc_code":"5792-7","official_name":"Urine Glucose","unit":"mg/dL"}},"ketones":{"type":"string","title":"Ketones","metadata":{"loinc_code":"5797-6","official_name":"Urine Ketones","unit":"mg/dL"}},"bilirubin":{"type":"string","title":"Bilirubin","metadata":{"loinc_code":"5802-4","official_name":"Urine Bilirubin","unit":"mg/dL"}},"rbc":{"type":"string","title":"Rbc","metadata":{"loinc_code":"58445-8","official_name":"Urine Red Blood Cells","unit":"/hpf"}},"pusCells":{"type":"string","title":"Puscells","metadata":{"loinc_code":"5821-4","official_name":"Urine White Blood Cells","unit":"/hpf"}},"epithelialCells":{"type":"string","title":"Epithelialcells","metadata":{"loinc_code":"58439-1","official_name":"Urine Epithelial Cells","unit":"/hpf"}},"casts":{"type":"string","title":"Casts","metadata":{"loinc_code":"5809-9","official_name":"Urine Casts","unit":"/lpf"}},"crystals":{"type":"string","title":"Crystals","metadata":{"loinc_code":"58441-7","official_name":"Urine Crystals","unit":"/hpf"}},"bacteria":{"type":"string","title":"Bacteria","metadata":{"loinc_code":"5822-2","official_name":"Urine Bacteria","unit":"/hpf"}}},"type":"object","required":["quantity","color","transparency","specificGravity","pH","protein","sugar","ketones","bilirubin","rbc","pusCells","epithelialCells","casts","crystals","bacteria"],"title":"UFRTemplate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}